requirements.txt
asttokens==2.4.1
build==1.2.1
CacheControl==0.14.0
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
crashtest==0.4.1
cvzone==1.6.1
cycler==0.10.0
Cython==3.0.10
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.8
dulwich==0.21.7
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
idna==2.10
imageio==2.34.1
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
imutils==0.5.4
installer==0.7.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.23.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
keyring==24.3.1
kiwisolver==1.4.5
lapx==0.5.7
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mkl==2021.4.0
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
mysql-connector==2.2.9
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
packaging==24.0
pafy==0.5.5
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.1
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.7.1
proglog==0.1.10
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.17.2
pyparsing==3.1.2
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.0.2
rapidfuzz==3.9.0
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.27
scipy==1.13.0
seaborn==0.13.2
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
stack-data==0.6.3
supervision==0.20.0
sv-ttk==2.6.0
sympy==1.12
tbb==2021.12.0
thop==0.1.1.post2209072238
tomlkit==0.12.5
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
trove-classifiers==2024.4.10
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.2
urllib3==1.26.18
virtualenv==20.26.1
wcwidth==0.2.13
youtube-dl==2021.12.17
zipp==3.18.1
